requirements.txt
# ─── Web 框架 ─────────────────────────────────────────────
aiohttp==3.9.5
aiofiles==23.2.1

# ─── 数据库 ──────────────────────────────────────────────
asyncpg==0.29.0           # PostgreSQL 异步驱动（运行时）
psycopg2-binary==2.9.9    # PostgreSQL 同步驱动（Alembic 迁移用）
alembic==1.13.1
sqlalchemy==2.0.29        # 仅供 Alembic 使用，不用 ORM

# ─── 缓存 + 队列 ──────────────────────────────────────────
redis[asyncio]==5.0.3     # aioredis 已并入 redis-py
rq==1.16.1                # Redis Queue（Worker 任务队列）

# ─── 向量数据库 ───────────────────────────────────────────
pymilvus==2.4.0

# ─── 认证 ────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# ─── AI / LLM ─────────────────────────────────────────────
openai==1.35.0            # 兼容 Qwen DashScope + DeepSeek OpenAI 接口

# ─── 文档解析 ─────────────────────────────────────────────
pdfplumber==0.11.0        # PDF 解析
openpyxl==3.1.4           # Excel 解析
python-docx==1.1.2        # Word 解析

# ─── 工具库 ───────────────────────────────────────────────
python-dotenv==1.0.1      # .env 加载
pydantic==2.7.0           # 数据校验
python-multipart==0.0.9   # 文件上传解析

# ─── 测试 ────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
